LPD-75170 refactor: move /ai-hub/ai-hub-web to /ai-hub-nexus/ai-hub-nexus-web
