+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay AI Hub Nexus Web
+Bundle-SymbolicName: com.liferay.ai.hub.nexus.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /ai-hub-nexus-web
+-dsannotations-options: inherit
