+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/build.gradle
@@ -1,0 +1,25 @@
+dependencies {
+	compileOnly group: "com.liferay.jakarta.portlet", name: "com.liferay.jakarta.portlet-api", version: "4.0.0"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.util.taglib", version: "default"
+	compileOnly group: "jakarta.servlet", name: "jakarta.servlet-api", version: "6.0.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
+	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:account:account-api")
+	compileOnly project(":apps:batch-engine:batch-engine-api")
+	compileOnly project(":apps:fragment:fragment-api")
+	compileOnly project(":apps:frontend-data-set:frontend-data-set-api")
+	compileOnly project(":apps:frontend-data-set:frontend-data-set-taglib")
+	compileOnly project(":apps:frontend-taglib:frontend-taglib")
+	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
+	compileOnly project(":apps:frontend-taglib:frontend-taglib-react")
+	compileOnly project(":apps:object:object-admin-rest-api")
+	compileOnly project(":apps:portal-workflow:portal-workflow-api")
+	compileOnly project(":apps:site:site-api")
+	compileOnly project(":core:petra:petra-function")
+	compileOnly project(":core:petra:petra-io")
+	compileOnly project(":core:petra:petra-lang")
+	compileOnly project(":core:petra:petra-string")
+	compileOnly project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-api")
+}
