+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/java/com/liferay/ai/hub/nexus/web/internal/constants/AIHubNexusFDSNames.java
@@ -1,0 +1,16 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.nexus.web.internal.constants;
+
+/**
+ * @author Feliphe Marinho
+ */
+public class AIHubNexusFDSNames {
+
+	public static final String AGENT_DEFINITIONS =
+		AIHubNexusWebConstants.BUNDLE_SYMBOLIC_NAME + "-agentDefinitions";
+
+}
