+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/java/com/liferay/ai/hub/nexus/web/internal/fragment/renderer/BaseFragmentRenderer.java
@@ -1,0 +1,73 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.nexus.web.internal.fragment.renderer;
+
+import com.liferay.ai.hub.nexus.web.internal.constants.AIHubNexusWebConstants;
+import com.liferay.fragment.renderer.FragmentRenderer;
+import com.liferay.fragment.renderer.FragmentRendererContext;
+
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.ServletContext;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Davyson Melo
+ * @author Feliphe Marinho
+ */
+public abstract class BaseFragmentRenderer<T> implements FragmentRenderer {
+
+	@Override
+	public boolean isSelectable(HttpServletRequest httpServletRequest) {
+		return false;
+	}
+
+	@Override
+	public void render(
+			FragmentRendererContext fragmentRendererContext,
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
+
+		try {
+			RequestDispatcher requestDispatcher =
+				servletContext.getRequestDispatcher(getJSPPath());
+
+			T displayContext = getDisplayContext(httpServletRequest);
+
+			if (displayContext != null) {
+				Class<?> clazz = displayContext.getClass();
+
+				httpServletRequest.setAttribute(
+					clazz.getName(), displayContext);
+			}
+
+			requestDispatcher.include(httpServletRequest, httpServletResponse);
+		}
+		catch (IOException | RuntimeException exception) {
+			throw exception;
+		}
+		catch (Exception exception) {
+			throw new IOException(exception);
+		}
+	}
+
+	protected T getDisplayContext(HttpServletRequest httpServletRequest) {
+		return null;
+	}
+
+	protected abstract String getJSPPath();
+
+	@Reference(
+		target = "(osgi.web.symbolicname=" + AIHubNexusWebConstants.BUNDLE_SYMBOLIC_NAME + ")"
+	)
+	protected ServletContext servletContext;
+
+}
