+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/resources/META-INF/resources/edit_agent_definition.jsp
@@ -1,0 +1,19 @@
+<%--
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+--%>
+
+<%@ include file="/init.jsp" %>
+
+<%
+EditAgentDefinitionDisplayContext editAgentDefinitionDisplayContext = (EditAgentDefinitionDisplayContext)request.getAttribute(EditAgentDefinitionDisplayContext.class.getName());
+%>
+
+<div>
+	<react:component
+		module="{AgentDefinitionForm} from ai-hub-nexus-web"
+		props="<%= editAgentDefinitionDisplayContext.getReactData() %>"
+	/>
+</div>
