+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/resources/META-INF/resources/js/agent_definition_form/components/ToolBar.tsx
