--- modules/dxp/apps/ai-hub/ai-hub-web/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay AI Hub Web
-Bundle-SymbolicName: com.liferay.ai.hub.web
-Bundle-Version: 1.0.2
-Web-ContextPath: /ai-hub-web
--dsannotations-options: inherit
