--- modules/dxp/apps/ai-hub/ai-hub-web/src/main/java/com/liferay/ai/hub/web/internal/constants/AIHubWebConstants.java
@@ -1,15 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.web.internal.constants;
-
-/**
- * @author Feliphe Marinho
- */
-public class AIHubWebConstants {
-
-	public static final String BUNDLE_SYMBOLIC_NAME = "com.liferay.ai.hub.web";
-
-}
