--- modules/dxp/apps/ai-hub/ai-hub-web/src/main/resources/META-INF/resources/tsconfig.json
@@ -1,53 +1,0 @@
-{
-	"@generated": "ea5005e318d058b2ac5d81d29b911af68107e8ab",
-	"@readonly": "** AUTO-GENERATED: DO NOT EDIT **",
-	"compilerOptions": {
-		"allowSyntheticDefaultImports": true,
-		"checkJs": false,
-		"composite": true,
-		"declarationDir": "../../../../../../../../../.tsc/types",
-		"emitDeclarationOnly": true,
-		"jsx": "react",
-		"module": "es2022",
-		"moduleResolution": "node",
-		"paths": {
-			"@liferay/object-js-components-web": [
-				"../../../../../../../../../apps/object/object-js-components-web/src/main/resources/META-INF/resources/index.ts"
-			],
-			"frontend-js-components-web": [
-				"../../../../../../../../../apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/index.ts"
-			],
-			"frontend-js-web": [
-				"../../../../../../../../../apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/main/index.d.ts"
-			],
-			"frontend-js-web/legacy": [
-				"../../../../../../../../../apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/legacy/index.ts"
-			]
-		},
-		"rootDir": "../../../../../../../../..",
-		"skipLibCheck": true,
-		"sourceMap": false,
-		"strict": true,
-		"target": "esnext",
-		"tsBuildInfoFile": "../../../../../../../../../.tsc/buildinfo/@liferay/ai-hub-web.tsbuildinfo",
-		"typeRoots": [
-			"../../../../../../../../../node_modules/@types"
-		]
-	},
-	"include": [
-		"**/*.ts",
-		"**/*.tsx",
-		"../../../../../../../../../global.d.ts"
-	],
-	"references": [
-		{
-			"path": "../../../../../../../../../apps/object/object-js-components-web/src/main/resources/META-INF/resources/tsconfig.json"
-		},
-		{
-			"path": "../../../../../../../../../apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/tsconfig.json"
-		},
-		{
-			"path": "../../../../../../../../../apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/tsconfig.json"
-		}
-	]
-}
